+++ Tests/Runtime/MediaStreamTest.cs
@@ -3,297 +3,11 @@ using UnityEngine.TestTools;
 using NUnit.Framework;
 using System.Linq;
 using System.Collections;
-using System.Collections.Generic;
-using NUnit.Framework.Constraints;
 
 namespace Unity.WebRTC.RuntimeTest
 {
-    class Ignore
-    {
-        static public void Pass<T>(T val)
-        {
-        }
-    }
-
     class MediaStreamTest
     {
-        void StatsTest(RTCStats stats)
-        {
-            switch (stats.Type)
-            {
-                case RTCStatsType.CandidatePair:
-                    var iceCandidatePairStats = stats as RTCIceCandidatePairStats;
-                    Assert.NotNull(iceCandidatePairStats);
-                    Assert.AreEqual(24, iceCandidatePairStats.Dict.Count);
-                    Assert.IsNotEmpty(iceCandidatePairStats.transportId);
-                    Assert.IsNotEmpty(iceCandidatePairStats.localCandidateId);
-                    Assert.IsNotEmpty(iceCandidatePairStats.remoteCandidateId);
-                    Assert.IsNotEmpty(iceCandidatePairStats.state);
-                    Ignore.Pass(iceCandidatePairStats.priority);
-                    Ignore.Pass(iceCandidatePairStats.nominated);
-                    Ignore.Pass(iceCandidatePairStats.writable);
-                    Ignore.Pass(iceCandidatePairStats.readable);
-                    Ignore.Pass(iceCandidatePairStats.bytesSent);
-                    Ignore.Pass(iceCandidatePairStats.bytesReceived);
-                    Ignore.Pass(iceCandidatePairStats.totalRoundTripTime);
-                    Ignore.Pass(iceCandidatePairStats.currentRoundTripTime);
-                    Ignore.Pass(iceCandidatePairStats.availableOutgoingBitrate);
-                    Ignore.Pass(iceCandidatePairStats.availableIncomingBitrate);
-                    Ignore.Pass(iceCandidatePairStats.consentRequestsReceived);
-                    Ignore.Pass(iceCandidatePairStats.consentRequestsSent);
-                    Ignore.Pass(iceCandidatePairStats.requestsReceived);
-                    Ignore.Pass(iceCandidatePairStats.requestsSent);
-                    Ignore.Pass(iceCandidatePairStats.responsesReceived);
-                    Ignore.Pass(iceCandidatePairStats.responsesSent);
-                    Ignore.Pass(iceCandidatePairStats.retransmissionsReceived);
-                    Ignore.Pass(iceCandidatePairStats.retransmissionsSent);
-                    Ignore.Pass(iceCandidatePairStats.consentResponsesReceived);
-                    Ignore.Pass(iceCandidatePairStats.consentResponsesSent);
-                    break;
-                case RTCStatsType.DataChannel:
-                    var dataChannelStats = stats as RTCDataChannelStats;
-                    Assert.NotNull(dataChannelStats);
-                    Assert.AreEqual(8, dataChannelStats.Dict.Count);
-                    Assert.IsNotEmpty(dataChannelStats.label);
-                    Assert.IsNotEmpty(dataChannelStats.state);
-                    Ignore.Pass(dataChannelStats.protocol);
-                    Ignore.Pass(dataChannelStats.messagesSent);
-                    Ignore.Pass(dataChannelStats.messagesReceived);
-                    Ignore.Pass(dataChannelStats.datachannelid);
-                    Ignore.Pass(dataChannelStats.bytesSent);
-                    Ignore.Pass(dataChannelStats.bytesReceived);
-                    break;
-                case RTCStatsType.LocalCandidate:
-                case RTCStatsType.RemoteCandidate:
-                    var candidateStats = stats as RTCIceCandidateStats;
-                    Assert.NotNull(candidateStats);
-                    Assert.AreEqual(11, candidateStats.Dict.Count);
-                    Assert.IsNotEmpty(candidateStats.protocol);
-                    Assert.IsNotEmpty(candidateStats.candidateType);
-                    Ignore.Pass(candidateStats.ip);
-                    Assert.IsNotEmpty(candidateStats.transportId);
-                    Ignore.Pass(candidateStats.url);
-                    Ignore.Pass(candidateStats.relayProtocol);
-                    Ignore.Pass(candidateStats.networkType);
-                    Ignore.Pass(candidateStats.priority);
-                    Ignore.Pass(candidateStats.deleted);
-                    Ignore.Pass(candidateStats.isRemote);
-                    Ignore.Pass(candidateStats.port);
-                    break;
-                case RTCStatsType.Certificate:
-                    var certificateStats = stats as RTCCertificateStats;
-                    Assert.NotNull(certificateStats);
-                    Assert.AreEqual(4, certificateStats.Dict.Count);
-                    Assert.IsNotEmpty(certificateStats.fingerprint);
-                    Assert.IsNotEmpty(certificateStats.fingerprintAlgorithm);
-                    Assert.IsNotEmpty(certificateStats.base64Certificate);
-                    Ignore.Pass(certificateStats.issuerCertificateId);
-                    break;
-                case RTCStatsType.Codec:
-                    var codecStats = stats as RTCCodecStats;
-                    Assert.NotNull(codecStats);
-                    Assert.AreEqual(5, codecStats.Dict.Count);
-                    Assert.IsNotEmpty(codecStats.mimeType);
-                    Ignore.Pass(codecStats.sdpFmtpLine);
-                    Ignore.Pass(codecStats.payloadType);
-                    Ignore.Pass(codecStats.clockRate);
-                    Ignore.Pass(codecStats.channels);
-                    break;
-                case RTCStatsType.InboundRtp:
-                    var inboundRtpStreamStats = stats as RTCInboundRTPStreamStats;
-                    Assert.NotNull(inboundRtpStreamStats);
-                    Assert.AreEqual(39, inboundRtpStreamStats.Dict.Count);
-                    Ignore.Pass(inboundRtpStreamStats.ssrc);
-                    Ignore.Pass(inboundRtpStreamStats.isRemote);
-                    Assert.IsNotEmpty(inboundRtpStreamStats.mediaType);
-                    Assert.IsNotEmpty(inboundRtpStreamStats.kind);
-                    Ignore.Pass(inboundRtpStreamStats.trackId);
-                    Assert.IsNotEmpty(inboundRtpStreamStats.transportId);
-                    Ignore.Pass(inboundRtpStreamStats.codecId);
-                    Ignore.Pass(inboundRtpStreamStats.firCount);
-                    Ignore.Pass(inboundRtpStreamStats.pliCount);
-                    Ignore.Pass(inboundRtpStreamStats.nackCount);
-                    Ignore.Pass(inboundRtpStreamStats.sliCount);
-                    Ignore.Pass(inboundRtpStreamStats.qpSum);
-                    Ignore.Pass(inboundRtpStreamStats.packetsReceived);
-                    Ignore.Pass(inboundRtpStreamStats.fecPacketsReceived);
-                    Ignore.Pass(inboundRtpStreamStats.fecPacketsDiscarded);
-                    Ignore.Pass(inboundRtpStreamStats.bytesReceived);
-                    Ignore.Pass(inboundRtpStreamStats.headerBytesReceived);
-                    Ignore.Pass(inboundRtpStreamStats.packetsLost);
-                    Ignore.Pass(inboundRtpStreamStats.lastPacketReceivedTimestamp);
-                    Ignore.Pass(inboundRtpStreamStats.jitter);
-                    Ignore.Pass(inboundRtpStreamStats.roundTripTime);
-                    Ignore.Pass(inboundRtpStreamStats.packetsDiscarded);
-                    Ignore.Pass(inboundRtpStreamStats.packetsRepaired);
-                    Ignore.Pass(inboundRtpStreamStats.burstPacketsLost);
-                    Ignore.Pass(inboundRtpStreamStats.burstPacketsDiscarded);
-                    Ignore.Pass(inboundRtpStreamStats.burstLossCount);
-                    Ignore.Pass(inboundRtpStreamStats.burstDiscardCount);
-                    Ignore.Pass(inboundRtpStreamStats.burstLossRate);
-                    Ignore.Pass(inboundRtpStreamStats.burstDiscardRate);
-                    Ignore.Pass(inboundRtpStreamStats.gapLossRate);
-                    Ignore.Pass(inboundRtpStreamStats.gapDiscardRate);
-                    Ignore.Pass(inboundRtpStreamStats.framesDecoded);
-                    Ignore.Pass(inboundRtpStreamStats.keyFramesDecoded);
-                    Ignore.Pass(inboundRtpStreamStats.totalDecodeTime);
-                    Ignore.Pass(inboundRtpStreamStats.totalInterFrameDelay);
-                    Ignore.Pass(inboundRtpStreamStats.totalSquaredInterFrameDelay);
-                    Ignore.Pass(inboundRtpStreamStats.contentType);
-                    Ignore.Pass(inboundRtpStreamStats.estimatedPlayoutTimestamp);
-                    Ignore.Pass(inboundRtpStreamStats.decoderImplementation);
-                    break;
-                case RTCStatsType.RemoteInboundRtp:
-                    var remoteInboundRtpStreamStats = stats as RTCRemoteInboundRtpStreamStats;
-                    Assert.NotNull(remoteInboundRtpStreamStats);
-                    Assert.AreEqual(8, remoteInboundRtpStreamStats.Dict.Count);
-                    Ignore.Pass(remoteInboundRtpStreamStats.ssrc);
-                    Assert.IsNotEmpty(remoteInboundRtpStreamStats.kind);
-                    Assert.IsNotEmpty(remoteInboundRtpStreamStats.transportId);
-                    Assert.IsNotEmpty(remoteInboundRtpStreamStats.codecId);
-                    Ignore.Pass(remoteInboundRtpStreamStats.packetsLost);
-                    Ignore.Pass(remoteInboundRtpStreamStats.jitter);
-                    Assert.IsNotEmpty(remoteInboundRtpStreamStats.localId);
-                    Ignore.Pass(remoteInboundRtpStreamStats.roundTripTime);
-                    break;
-                case RTCStatsType.Transport:
-                    var transportStats = stats as RTCTransportStats;
-                    Assert.NotNull(transportStats);
-                    Assert.AreEqual(11, transportStats.Dict.Count);
-                    Ignore.Pass(transportStats.bytesSent);
-                    Ignore.Pass(transportStats.bytesReceived);
-                    Ignore.Pass(transportStats.rtcpTransportStatsId);
-                    Ignore.Pass(transportStats.dtlsState);
-                    Ignore.Pass(transportStats.selectedCandidatePairId);
-                    Ignore.Pass(transportStats.localCertificateId);
-                    Ignore.Pass(transportStats.remoteCertificateId);
-                    Ignore.Pass(transportStats.tlsVersion);
-                    Ignore.Pass(transportStats.dtlsCipher);
-                    Ignore.Pass(transportStats.srtpCipher);
-                    Ignore.Pass(transportStats.selectedCandidatePairChanges);
-                    break;
-                case RTCStatsType.RemoteOutboundRtp:
-                case RTCStatsType.OutboundRtp:
-                    var outboundRtpStreamStats = stats as RTCOutboundRTPStreamStats;
-                    Assert.NotNull(outboundRtpStreamStats);
-                    Assert.AreEqual(35, outboundRtpStreamStats.Dict.Count);
-                    Ignore.Pass(outboundRtpStreamStats.ssrc);
-                    Ignore.Pass(outboundRtpStreamStats.isRemote);
-                    Assert.IsNotEmpty(outboundRtpStreamStats.mediaType);
-                    Assert.IsNotEmpty(outboundRtpStreamStats.kind);
-                    Ignore.Pass(outboundRtpStreamStats.trackId);
-                    Assert.IsNotEmpty(outboundRtpStreamStats.transportId);
-                    Ignore.Pass(outboundRtpStreamStats.codecId);
-                    Ignore.Pass(outboundRtpStreamStats.firCount);
-                    Ignore.Pass(outboundRtpStreamStats.pliCount);
-                    Ignore.Pass(outboundRtpStreamStats.nackCount);
-                    Ignore.Pass(outboundRtpStreamStats.sliCount);
-                    Ignore.Pass(outboundRtpStreamStats.qpSum);
-                    Ignore.Pass(outboundRtpStreamStats.mediaSourceId);
-                    Ignore.Pass(outboundRtpStreamStats.remoteId);
-                    Ignore.Pass(outboundRtpStreamStats.rid);
-                    Ignore.Pass(outboundRtpStreamStats.packetsSent);
-                    Ignore.Pass(outboundRtpStreamStats.retransmittedPacketsSent);
-                    Ignore.Pass(outboundRtpStreamStats.bytesSent);
-                    Ignore.Pass(outboundRtpStreamStats.headerBytesSent);
-                    Ignore.Pass(outboundRtpStreamStats.retransmittedBytesSent);
-                    Ignore.Pass(outboundRtpStreamStats.targetBitrate);
-                    Ignore.Pass(outboundRtpStreamStats.framesEncoded);
-                    Ignore.Pass(outboundRtpStreamStats.keyFramesEncoded);
-                    Ignore.Pass(outboundRtpStreamStats.totalEncodeTime);
-                    Ignore.Pass(outboundRtpStreamStats.totalEncodedBytesTarget);
-                    Ignore.Pass(outboundRtpStreamStats.frameWidth);
-                    Ignore.Pass(outboundRtpStreamStats.frameHeight);
-                    Ignore.Pass(outboundRtpStreamStats.framesPerSecond);
-                    Ignore.Pass(outboundRtpStreamStats.framesSent);
-                    Ignore.Pass(outboundRtpStreamStats.hugeFramesSent);
-                    Ignore.Pass(outboundRtpStreamStats.totalPacketSendDelay);
-                    Ignore.Pass(outboundRtpStreamStats.qualityLimitationReason);
-                    Ignore.Pass(outboundRtpStreamStats.qualityLimitationResolutionChanges);
-                    Ignore.Pass(outboundRtpStreamStats.contentType);
-                    Ignore.Pass(outboundRtpStreamStats.encoderImplementation);
-                    break;
-                case RTCStatsType.MediaSource:
-                    var mediaSourceStats = stats as RTCMediaSourceStats;
-                    Assert.NotNull(mediaSourceStats);
-                    Assert.AreEqual(6, mediaSourceStats.Dict.Count);
-                    Assert.IsNotEmpty(mediaSourceStats.trackIdentifier);
-                    Assert.IsNotEmpty(mediaSourceStats.kind);
-                    Ignore.Pass(mediaSourceStats.width);
-                    Ignore.Pass(mediaSourceStats.height);
-                    Ignore.Pass(mediaSourceStats.frames);
-                    Ignore.Pass(mediaSourceStats.framesPerSecond);
-
-                    break;
-                case RTCStatsType.Track:
-                    var mediaStreamTrackStats = stats as RTCMediaStreamTrackStats;
-                    Assert.NotNull(mediaStreamTrackStats);
-                    Assert.AreEqual(42, mediaStreamTrackStats.Dict.Count);
-                    Ignore.Pass(mediaStreamTrackStats.trackIdentifier);
-                    Ignore.Pass(mediaStreamTrackStats.mediaSourceId);
-                    Ignore.Pass(mediaStreamTrackStats.remoteSource);
-                    Ignore.Pass(mediaStreamTrackStats.ended);
-                    Ignore.Pass(mediaStreamTrackStats.detached);
-                    Ignore.Pass(mediaStreamTrackStats.kind);
-                    Ignore.Pass(mediaStreamTrackStats.jitterBufferDelay);
-                    Ignore.Pass(mediaStreamTrackStats.jitterBufferEmittedCount);
-                    Ignore.Pass(mediaStreamTrackStats.frameWidth);
-                    Ignore.Pass(mediaStreamTrackStats.frameHeight);
-                    Ignore.Pass(mediaStreamTrackStats.framesPerSecond);
-                    Ignore.Pass(mediaStreamTrackStats.framesSent);
-                    Ignore.Pass(mediaStreamTrackStats.hugeFramesSent);
-                    Ignore.Pass(mediaStreamTrackStats.framesReceived);
-                    Ignore.Pass(mediaStreamTrackStats.framesDecoded);
-                    Ignore.Pass(mediaStreamTrackStats.framesDropped);
-                    Ignore.Pass(mediaStreamTrackStats.framesCorrupted);
-                    Ignore.Pass(mediaStreamTrackStats.partialFramesLost);
-                    Ignore.Pass(mediaStreamTrackStats.fullFramesLost);
-                    Ignore.Pass(mediaStreamTrackStats.audioLevel);
-                    Ignore.Pass(mediaStreamTrackStats.totalAudioEnergy);
-                    Ignore.Pass(mediaStreamTrackStats.echoReturnLoss);
-                    Ignore.Pass(mediaStreamTrackStats.echoReturnLossEnhancement);
-                    Ignore.Pass(mediaStreamTrackStats.totalSamplesReceived);
-                    Ignore.Pass(mediaStreamTrackStats.totalSamplesDuration);
-                    Ignore.Pass(mediaStreamTrackStats.concealedSamples);
-                    Ignore.Pass(mediaStreamTrackStats.silentConcealedSamples);
-                    Ignore.Pass(mediaStreamTrackStats.concealmentEvents);
-                    Ignore.Pass(mediaStreamTrackStats.insertedSamplesForDeceleration);
-                    Ignore.Pass(mediaStreamTrackStats.removedSamplesForAcceleration);
-                    Ignore.Pass(mediaStreamTrackStats.jitterBufferFlushes);
-                    Ignore.Pass(mediaStreamTrackStats.delayedPacketOutageSamples);
-                    Ignore.Pass(mediaStreamTrackStats.relativePacketArrivalDelay);
-                    Ignore.Pass(mediaStreamTrackStats.jitterBufferTargetDelay);
-                    Ignore.Pass(mediaStreamTrackStats.interruptionCount);
-                    Ignore.Pass(mediaStreamTrackStats.totalInterruptionDuration);
-                    Ignore.Pass(mediaStreamTrackStats.freezeCount);
-                    Ignore.Pass(mediaStreamTrackStats.pauseCount);
-                    Ignore.Pass(mediaStreamTrackStats.totalFreezesDuration);
-                    Ignore.Pass(mediaStreamTrackStats.totalPausesDuration);
-                    Ignore.Pass(mediaStreamTrackStats.totalFramesDuration);
-                    Ignore.Pass(mediaStreamTrackStats.sumOfSquaredFramesDuration);
-                    break;
-                case RTCStatsType.Stream:
-                    var mediaStreamStats = stats as RTCMediaStreamStats;
-                    Assert.NotNull(mediaStreamStats);
-                    Assert.AreEqual(2, mediaStreamStats.Dict.Count);
-                    Ignore.Pass(mediaStreamStats.streamIdentifier);
-                    Ignore.Pass(mediaStreamStats.trackIds);
-                    break;
-                case RTCStatsType.PeerConnection:
-                    var peerConnectionStats = stats as RTCPeerConnectionStats;
-                    Assert.NotNull(peerConnectionStats);
-                    Assert.AreEqual(2, peerConnectionStats.Dict.Count);
-                    Ignore.Pass(peerConnectionStats.dataChannelsOpened);
-                    Ignore.Pass(peerConnectionStats.dataChannelsClosed);
-                    break;
-                default:
-                    Debug.LogWarning(stats.Type);
-                    break;
-            }
-        }
-
-
-
         [SetUp]
         public void SetUp()
         {
@@ -484,7 +198,7 @@ namespace Unity.WebRTC.RuntimeTest
                 new RTCIceServer {urls = new[] {"stun:stun.l.google.com:19302"}}
             };
             var audioStream = Audio.CaptureStream();
-            var test = new MonoBehaviourTest<SignalingPeersTest>();
+            var test = new MonoBehaviourTest<SignalingPeers>();
             test.component.SetStream(audioStream);
             yield return test;
             test.component.Dispose();
@@ -504,54 +218,10 @@ namespace Unity.WebRTC.RuntimeTest
             var videoStream = cam.CaptureStream(1280, 720, 1000000);
             yield return new WaitForSeconds(0.1f);
 
-            var test = new MonoBehaviourTest<SignalingPeersTest>();
+            var test = new MonoBehaviourTest<SignalingPeers>();
             test.component.SetStream(videoStream);
             yield return test;
             yield return new WaitForSeconds(0.1f);
-            test.component.Dispose();
-            foreach (var track in videoStream.GetTracks())
-            {
-                track.Dispose();
-            }
-            videoStream.Dispose();
-            Object.DestroyImmediate(camObj);
-        }
-
-        [UnityTest]
-        [Timeout(5000)]
-        public IEnumerator PeerConnectionGetStats()
-        {
-            var camObj = new GameObject("Camera");
-            var cam = camObj.AddComponent<Camera>();
-            var videoStream = cam.CaptureStream(1280, 720, 1000000);
-            yield return new WaitForSeconds(0.1f);
-
-            var test = new MonoBehaviourTest<SignalingPeersTest>();
-            test.component.SetStream(videoStream);
-            yield return test;
-            test.component.CoroutineWebRTCUpdate();
-            yield return new WaitForSeconds(0.1f);
-            var op = test.component.GetPeerStats();
-            yield return op;
-            Assert.True(op.IsDone);
-            Assert.IsNotEmpty(op.Value.Stats);
-            Assert.IsNotEmpty(op.Value.Stats.Keys);
-            Assert.IsNotEmpty(op.Value.Stats.Values);
-            Assert.Greater(op.Value.Stats.Count, 0);
-
-            foreach (RTCStats stats in op.Value.Stats.Values)
-            {
-                Assert.NotNull(stats);
-                Assert.Greater(stats.Timestamp, 0);
-                Assert.IsNotEmpty(stats.Id);
-                foreach(var pair in stats.Dict)
-                {
-                    Assert.IsNotEmpty(pair.Key);
-                    Assert.NotNull(pair.Value);
-                }
-                StatsTest(stats);
-            }
-            op.Value.Dispose();
             test.component.Dispose();
             foreach (var track in videoStream.GetTracks())
             {
@@ -570,10 +240,10 @@ namespace Unity.WebRTC.RuntimeTest
             var videoStream = cam.CaptureStream(1280, 720, 1000000);
             yield return new WaitForSeconds(0.1f);
 
-            var test = new MonoBehaviourTest<SignalingPeersTest>();
+            var test = new MonoBehaviourTest<SignalingPeers>();
             test.component.SetStream(videoStream);
             yield return test;
-            test.component.CoroutineWebRTCUpdate();
+            test.component.CoroutineUpdate();
             yield return new WaitForSeconds(0.1f);
             var op = test.component.GetSenderStats(0);
             yield return op;
@@ -591,7 +261,7 @@ namespace Unity.WebRTC.RuntimeTest
                     Assert.IsNotEmpty(pair.Key);
                     Assert.NotNull(pair.Value);
                 }
-                StatsTest(stats);
+                StatsCheck.Test(stats);
             }
 
             op.Value.Dispose();
@@ -613,9 +283,11 @@ namespace Unity.WebRTC.RuntimeTest
             var videoStream = cam.CaptureStream(1280, 720, 1000000);
             yield return new WaitForSeconds(0.1f);
 
-            var test = new MonoBehaviourTest<SignalingPeersTest>();
+            var test = new MonoBehaviourTest<SignalingPeers>();
             test.component.SetStream(videoStream);
             yield return test;
+            test.component.CoroutineUpdate();
+            yield return new WaitForSeconds(0.1f);
             var op = test.component.GetReceiverStats(0);
             yield return op;
             Assert.True(op.IsDone);
@@ -632,7 +304,7 @@ namespace Unity.WebRTC.RuntimeTest
                     Assert.IsNotEmpty(pair.Key);
                     Assert.NotNull(pair.Value);
                 }
-                StatsTest(stats);
+                StatsCheck.Test(stats);
             }
             test.component.Dispose();
             foreach (var track in videoStream.GetTracks())
@@ -664,10 +336,10 @@ namespace Unity.WebRTC.RuntimeTest
             var videoStream = cam.CaptureStream(1280, 720, 1000000);
             yield return new WaitForSeconds(0.1f);
 
-            var test = new MonoBehaviourTest<SignalingPeersTest>();
+            var test = new MonoBehaviourTest<SignalingPeers>();
             test.component.SetStream(videoStream);
             yield return test;
-            test.component.CoroutineWebRTCUpdate();
+            test.component.CoroutineUpdate();
             yield return new WaitForSeconds(0.1f);
 
             var senders = test.component.GetPeer1Senders();
@@ -692,143 +364,6 @@ namespace Unity.WebRTC.RuntimeTest
             }
             videoStream.Dispose();
             Object.DestroyImmediate(camObj);
-        }
-
-        public class SignalingPeersTest : MonoBehaviour, IMonoBehaviourTest
-        {
-            private bool m_isFinished;
-            private MediaStream m_stream;
-            RTCPeerConnection peer1;
-            RTCPeerConnection peer2;
-            RTCDataChannel dataChannel;
-
-            public bool IsTestFinished
-            {
-                get { return m_isFinished; }
-            }
-
-            public void SetStream(MediaStream stream)
-            {
-                m_stream = stream;
-            }
-
-            public RTCStatsReportAsyncOperation GetPeerStats()
-            {
-                return peer1.GetStats();
-            }
-
-            public RTCStatsReportAsyncOperation GetSenderStats(int index)
-            {
-                return GetPeer1Senders().ElementAt(index).GetStats();
-            }
-
-            public RTCStatsReportAsyncOperation GetReceiverStats(int index)
-            {
-                return GetPeer1Receivers().ElementAt(index).GetStats();
-            }
-
-            IEnumerator Start()
-            {
-                RTCConfiguration config = default;
-                config.iceServers = new[]
-                {
-                    new RTCIceServer {urls = new[] {"stun:stun.l.google.com:19302"}}
-                };
-
-                peer1 = new RTCPeerConnection(ref config);
-                peer2 = new RTCPeerConnection(ref config);
-                RTCDataChannelInit conf = new RTCDataChannelInit(true);
-                dataChannel = peer1.CreateDataChannel("data", ref conf);
-
-                peer1.OnIceCandidate = candidate =>
-                {
-                    Assert.NotNull(candidate);
-                    Assert.NotNull(candidate.candidate);
-                    peer2.AddIceCandidate(ref candidate);
-                };
-                peer2.OnIceCandidate = candidate =>
-                {
-                    Assert.NotNull(candidate);
-                    Assert.NotNull(candidate.candidate);
-                    peer1.AddIceCandidate(ref candidate);
-                };
-                peer2.OnTrack = e =>
-                {
-                    Assert.NotNull(e);
-                    Assert.NotNull(e.Track);
-                    Assert.NotNull(e.Receiver);
-                    Assert.NotNull(e.Transceiver);
-                    peer2.AddTrack(e.Track);
-                };
-
-                foreach (var track in m_stream.GetTracks())
-                {
-                    peer1.AddTrack(track, m_stream);
-                }
-
-                RTCOfferOptions options1 = default;
-                RTCAnswerOptions options2 = default;
-                var op1 = peer1.CreateOffer(ref options1);
-                yield return op1;
-                Assert.False(op1.IsError);
-                var desc = op1.Desc;
-                var op2 = peer1.SetLocalDescription(ref desc);
-                yield return op2;
-                Assert.False(op2.IsError);
-                var op3 = peer2.SetRemoteDescription(ref desc);
-                yield return op3;
-                Assert.False(op3.IsError);
-                var op4 = peer2.CreateAnswer(ref options2);
-                yield return op4;
-                Assert.False(op4.IsError);
-                desc = op4.Desc;
-                var op5 = peer2.SetLocalDescription(ref desc);
-                yield return op5;
-                Assert.False(op5.IsError);
-                var op6 = peer1.SetRemoteDescription(ref desc);
-                yield return op6;
-                Assert.False(op6.IsError);
-
-                var op7 = new WaitUntilWithTimeout(() =>
-                    peer1.IceConnectionState == RTCIceConnectionState.Connected ||
-                    peer1.IceConnectionState == RTCIceConnectionState.Completed, 5000);
-                yield return op7;
-                Assert.True(op7.IsCompleted);
-
-                var op8 = new WaitUntilWithTimeout(() =>
-                    peer2.IceConnectionState == RTCIceConnectionState.Connected ||
-                    peer2.IceConnectionState == RTCIceConnectionState.Completed, 5000);
-                yield return op8;
-                Assert.True(op8.IsCompleted);
-
-                var op9 = new WaitUntilWithTimeout(() => GetPeer1Senders().Any(), 5000);
-                yield return op9;
-                Assert.True(op9.IsCompleted);
-
-                m_isFinished = true;
-            }
-
-            public Coroutine CoroutineWebRTCUpdate()
-            {
-                return StartCoroutine(WebRTC.Update());
-            }
-
-            public IEnumerable<RTCRtpSender> GetPeer1Senders()
-            {
-                return peer1.GetSenders();
-            }
-
-            public IEnumerable<RTCRtpReceiver> GetPeer1Receivers()
-            {
-                return peer2.GetReceivers();
-            }
-
-            public void Dispose()
-            {
-                dataChannel?.Dispose();
-                peer1?.Close();
-                peer2?.Close();
-            }
         }
     }
 }
