+++ Tests/Runtime/SignalingPeers.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 234169ccc792f0f49a51f97f98428304
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
