+++ Tests/Runtime/StatsCheck.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 58cbe7a1d11171c43898ba79be53d6c9
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
